Release candidate 1: map widget, plots, export and general implementions
Map Widget
- map tips accuracy improved
- current layer logic improvements
- contours accuracy improved
- contours: progress interface
- edited Near Sided projection.
- vector layers management improvements
- vector plots: vector scale (including persistence and GUI),

Plots
- new Z axis plot titles & other minor improvements.
- title z for zfxy  (incomplete)
- both editors and control panels improvements

Operations
- export dialog and new projection: minor corrections.

General
- Brat Manual in main menu (incomplete)
- progress Interface
- new application options (map related)
- other minor corrections.
- build adjustments to all systems

+++ new-gui/Common/GUI/BasicLogShell.cpp
@@ -30,7 +30,7 @@ size_t CBasicLogShell::smMaxMsgsPerSec = smDefaultMaxMsgsPerSec;
 
 QGroupBox* CBasicLogShell::CreateDebugWidgets( const std::vector<std::string> &level_names )
 {
-	assert__( level_names.size() <= mSeverityToColorTable.size() && level_names.size() <= mSeverityToPromptTable.size() );
+	assert__( (int)level_names.size() <= mSeverityToColorTable.size() && (int)level_names.size() <= mSeverityToPromptTable.size() );
 
 	// create widgets
